break out OIDC login behavior into helper
